requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
alembic==1.14.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fs==2.4.16
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.18.6
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyotp==2.9.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
